change insecure default secrets in .env
This is to have some secure defaults for secrets
in general. Atleast a bit secure. This will make
development a bit harder but it will also instill
a good habit of not using insecure defaults, or
insecure secrets in general, especially for
serious projects. This is not a big deal, but just
a small step to make it a bit secure by default.
This will instill a good sense of security in
developers, especially beginners, who might not
know about the importance of secure secrets.

Signed-off-by: Karuppiah Natarajan <karuppiah7890@gmail.com>